feat(common): restructure project to move Prisma and config files to common directory; update imports accordingly
