+++ src/auth/constants/jwt.constants.ts
@@ -1,4 +1,4 @@
-import { envs } from 'src/config/envs';
+import { envs } from 'src/common/config/envs';
 
 export const jwtConstants = {
   secret: envs.jwtSecret,
