+++ src/common/helpers/helper.pagination.ts
@@ -1,4 +1,4 @@
-import { PrismaService } from 'src/prisma/prisma.service';
+import { PrismaService } from 'src/common/prisma/prisma.service';
 
 interface PaginationParams<T> {
   prisma: PrismaService;
