+++ src/course/dto/paginacion-course.ts
@@ -1,7 +1,7 @@
 import { IsEnum, IsOptional } from 'class-validator';
 import { PaginationDto } from 'src/common/dto/pagination.dto';
 import { CourseStatus } from '@prisma/client';
-import { CourseStatusList } from 'src/enums/course-status.enum';
+import { CourseStatusList } from 'src/common/enums/course-status.enum';
 
 export class CoursePaginationDto extends PaginationDto {
   @IsOptional()
