+++ src/enrollment/enrollment.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 import { EnrollmentService } from './enrollment.service';
 import { EnrollmentController } from './enrollment.controller';
-import { PrismaModule } from 'src/prisma/prisma.module';
+import { PrismaModule } from 'src/common/prisma/prisma.module';
 import { UsersModule } from 'src/user/users.module';
 import { UserProfileModule } from 'src/user-profile/user-profile.module';
