+++ src/user-profile/user-profile.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 import { UserProfileService } from './user-profile.service';
 import { UserProfileController } from './user-profile.controller';
-import { PrismaModule } from 'src/prisma/prisma.module';
+import { PrismaModule } from 'src/common/prisma/prisma.module';
 import { UsersModule } from 'src/user/users.module';
 
 @Module({
