+++ src/user/dto/user-paginacion.dto.ts
@@ -1,6 +1,6 @@
 import { IsEnum, IsOptional } from "class-validator";
 import { PaginationDto } from "src/common/dto/pagination.dto";
-import { UserStatusList} from "../../enums/user-status.enum";
+import { UserStatusList} from "../../common/enums/user-status.enum";
 import { UserStatus } from "@prisma/client";
 
 export class UserPaginationDto extends PaginationDto{
